Add PasswordSettingInput
